chore: split protocol.py into own folder/files
